requirements.txt
Django==5.2.*
djangorestframework
django-tailwind-cli
django-debug-toolbar
django-unfold
django-cotton


Scrapy
python-decouple
